split lodManager to another package
