+++ examples/room/lod/lodmanager/util.go
@@ -1,0 +1,8 @@
+package lodmanager
+
+func abs(x int32) int32 {
+	if x < 0 {
+		return -x
+	}
+	return x
+}
