+++ examples/room/lod/servers/udp/main.go
@@ -3,9 +3,21 @@
 package main
 
 import (
+	"github.com/Diarkis/diarkis/derror"
 	"github.com/Diarkis/diarkis/diarkisexec"
+	"github.com/Diarkis/diarkis/room"
+	"github.com/Diarkis/diarkis/server"
+	"github.com/Diarkis/diarkis/user"
 
+	"github.com/Diarkis/diarkis-server-template/examples/room/lod/lodmanager"
 	proom "github.com/Diarkis/diarkis-server-template/examples/room/lod/puffer/go/room"
+)
+
+var (
+	SyncIntervalForNearby int32 = 16
+	SyncIntervalForFar    int32 = 2000
+	MaxDistanceForNearby  int32 = 10000
+	MaxDistanceForFar     int32 = 40000
 )
 
 func main() {
@@ -20,5 +32,47 @@ func main() {
 	diarkisexec.SetServerCommandHandler(proom.BroadcastLoDVer, proom.BroadcastLoDCmd, handleRoomBroadcastLoD)
 	diarkisexec.SetServerCommandHandler(proom.GetLoDInfoVer, proom.GetLoDInfoCmd, handleRoomGetLoDInfo)
 
+	room.SetOnRoomDiscard(func(roomID string) {
+		lodmanager.RemoveRoomManager(roomID)
+	})
+
 	diarkisexec.StartDiarkis()
+}
+
+func handleRoomBroadcastLoD(ver uint8, cmd uint16, payload []byte, userData *user.User, next func(error)) {
+
+	proto := proom.NewBroadcastLoD()
+	err := proto.Unpack(payload)
+	if err != nil {
+		userData.ServerRespond(derror.ErrData(err.Error(), derror.InvalidParameter(0)), ver, cmd, server.Bad, true)
+		next(err)
+		return
+	}
+
+	roomID := room.GetRoomID(userData)
+	if roomID == "" {
+		userData.ServerRespond(derror.ErrData("Not in the room", derror.NotAllowed(0)), ver, cmd, server.Bad, true)
+		return
+	}
+
+	manager := lodmanager.GetRoomManager(roomID)
+	if manager == nil {
+		manager = lodmanager.NewManager(ver, cmd, SyncIntervalForNearby, SyncIntervalForFar, MaxDistanceForNearby, MaxDistanceForFar)
+		lodmanager.SetRoomManager(roomID, manager)
+	}
+	manager.AddUserEntity(userData.SID, proto.X, proto.Y, proto.Payload)
+	userData.ServerRespond(nil, ver, cmd, server.Ok, true)
+	next(nil)
+}
+
+func handleRoomGetLoDInfo(ver uint8, cmd uint16, payload []byte, userData *user.User, next func(error)) {
+	// TODO: Get the LoD information from configuration
+	lodInfo := proom.NewGetLoDInfoResponse()
+	lodInfo.SyncIntervalForNearby = int32(SyncIntervalForNearby)
+	lodInfo.SyncIntervalForFar = int32(SyncIntervalForFar)
+	lodInfo.MaxDistanceForNearby = int32(MaxDistanceForNearby)
+	lodInfo.MaxDistanceForFar = int32(MaxDistanceForFar)
+
+	userData.ServerRespond(lodInfo.Pack(), ver, cmd, server.Ok, true)
+	next(nil)
 }
